+++ core/runner/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nanoservice/runner
 
+## 0.1.10
+
+### Patch Changes
+
+- Fixed prometheus metrics
+
 ## 0.1.9
 
 ### Patch Changes